// Code/src/app/shared/model/GoddownOwnerDetailsModel.ts
export class GoddownOwnerDetailsModel {
    id: number;
    goddownOwnerCode: string;
    branch: number;
    goddown: number;
    ownerName: string;
    address: string;
    city: number;
    pinCode: string;
    state: number;
    stdCode: string;
    phoneNo: number;
    mobileNo: number;
    emailId: number;
    sqft: string;
    ratePerSqft: string;
    rentAmount: number;
    gst: number;
    totalRent: number;
    tds: number;
    netRent: number;
    pan: string;
    gstin: string;
    bank: number;
    ifsc: number;
    a_cno: string;
    chequeInTheNameOf: string;
    nextVersionDate: Date;
    holdPaymentInstruction: boolean;
    reamrk: string;
    isRented: boolean;
    rentPaymentType: string;
    securityDeposit: string;
    agreementStartDate: Date;
    agreementEndDate: Date;
    noticePeriodInDays: string;
    isActive: boolean;
}